requirements.txt
fastapi
uvicorn
pandas
python-multipart
pydantic
openai
python-dotenv
resend
